added copyrights correctly
