+++ components/org.wso2.carbon.identity.application.authenticator.biometric.endpoint/src/main/java/org/wso2/carbon/identity/application/authenticator/biometric/endpoint/servlet/BiometricServlet.java
@@ -1,28 +1,30 @@
 /*
- *  Copyright (c) 2019 , WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2019, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
- *  WSO2 Inc. licenses this file to you under the Apache License,
+ * WSO2 Inc. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
  * in compliance with the License.
- *  *  You may obtain a copy of the License at
- *  *
- *  *  http://www.apache.org/licenses/LICENSE-2.0
- *  *
- *  *  Unless required by applicable law or agreed to in writing,
- *  *  software distributed under the License is distributed on an
- *  *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- *  *  KIND, either express or implied.  See the License for the
- *  *  specific language governing permissions and limitations
- *  *  under the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ *
  */
 
-package org.wso2.carbon.identity.sso.saml.servlet;
+package org.wso2.carbon.identity.application.authenticator.biometric.endpoint.servlet;
 
 import com.google.gson.Gson;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
-import org.wso2.carbon.identity.sso.saml.javascript.flow.WaitStatusResponse;
-import org.wso2.carbon.identity.sso.saml.model.WaitStatus;
+import org.wso2.carbon.identity.application.authenticator.biometric.endpoint.model.WaitStatus;
+import org.wso2.carbon.identity.application.authenticator.biometric.endpoint.javascript.flow.WaitStatusResponse;
+
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.util.HashMap;
