+++ components/org.wso2.carbon.identity.application.authenticator.biometric/src/main/java/org/wso2/carbon/identity/application/authenticator/biometric/dao/BiometricDAO.java
@@ -17,23 +17,24 @@
  *
  */
 
+package org.wso2.carbon.identity.application.authenticator.biometric.dao;
 
-.loader {
-    border: 16px solid #f3f3f3; /* Light grey */
-    border-top: 16px solid #3498db; /* Blue */
-    border-radius: 50%;
-    width: 120px;
-    height: 120px;
-    animation: spin 2s linear infinite;
-    margin-left: auto;
-    margin-right: auto;
-}
+/**
+ * performs DAO operations related to Biometric device store.
+ */
+public interface BiometricDAO {
 
-@keyframes spin {
-    0% {
-        transform: rotate(0deg);
-    }
-    100% {
-        transform: rotate(360deg);
-    }
+
+    void addDeviceID(String username, String deviceID);
+
+
+    void updateDeviceID(String username, String deviceID);
+
+
+    void deleteDeviceID(String username);
+
+
+    String getDeviceID(String username);
+
+
 }
