+++ components/org.wso2.carbon.identity.application.authenticator.biometric/src/main/java/org/wso2/carbon/identity/application/authenticator/biometric/notification/handler/PushNotificationSender.java
@@ -17,23 +17,13 @@
  *
  */
 
+package org.wso2.carbon.identity.application.authenticator.biometric.notification.handler;
 
-.loader {
-    border: 16px solid #f3f3f3; /* Light grey */
-    border-top: 16px solid #3498db; /* Blue */
-    border-radius: 50%;
-    width: 120px;
-    height: 120px;
-    animation: spin 2s linear infinite;
-    margin-left: auto;
-    margin-right: auto;
-}
+/**
+ * Handles the sending of push notifications to specific device IDs.
+ */
+public interface PushNotificationSender {
 
-@keyframes spin {
-    0% {
-        transform: rotate(0deg);
-    }
-    100% {
-        transform: rotate(360deg);
-    }
+    void sendPushNotification(String deviceId, String serverKey,
+                              String message, String randomUUIDString, String sessionDataKey);
 }
